+++ src/cqterrain/tile/__init__.py
@@ -18,5 +18,6 @@ from .carton import carton
 from .carton2 import carton2
 from .truchet_triangle import truchet_triangle
 from .truchet_circle import truchet_circle
+from .dwarfStar import dwarf_star
 
 from .tile_plate import TilesPlate